refactor(cli-config): split cli/config.rs 1080 LOC en 10 sous-modules (#11)
Split `src/cli/config.rs` (1080 LOC god file) en 10 sous-modules par
domaine TOML. Ancien fichier renomme vers `config/mod.rs` puis allege
en module fin de re-exports (52 LOC).

Modules extraits :
- budget.rs        (22 LOC) BudgetConfig
- cache.rs         (49 LOC) CacheConfig
- security.rs     (216 LOC) SecurityConfig, ComplianceConfig, EnforcementMode, TeeConfig, FipsConfig
- server.rs       (140 LOC) ServerConfig, TimeoutConfig, TlsConfig, AcmeConfig
- telemetry.rs     (89 LOC) TracingConfig, OtelConfig
- user.rs          (32 LOC) UserConfig, PresetConfig
- routing.rs      (217 LOC) RouterConfig, PromptRule, ModelConfig, ModelMapping,
                            ModelStrategy, FanOutMode, FanOutConfig, TierConfig,
                            TierMatchCondition, ProjectConfig, ProjectRouterOverlay
- providers.rs    (149 LOC) ProviderConfig, AuthType, PoolConfig, PoolStrategy
- reliability.rs  (191 LOC) CircuitBreakerProviderConfig, HealthCheckProviderConfig,
                            parse_duration
- harness.rs       (22 LOC) HarnessConfig (cfg(feature = "harness"))

Tous les fichiers sont <= 220 LOC (seuil cible 300 LOC).

Move pur, aucune logique change, 0 diff comportemental. API publique
inchangee : `crate::cli::*` re-exporte via `config::*`. `default_true()`
partage entre security.rs et telemetry.rs est centralise dans mod.rs
comme `pub(crate) fn` (resolu via `super::default_true`).

Les doc-links `crate::cli::config::parse_duration` sont mis a jour vers
le chemin canonique public `crate::cli::parse_duration`.

Validation locale : cargo fmt + cargo clippy --all-features --all-targets
-D warnings + cargo test --all-features --lib (958 passed) + cargo test
--doc (19 passed). Commit pousse via gh api a cause d'une anomalie
ghost-reset dans le worktree (memory feedback_commis_push_ghosting).
